ADC is now internally buffered and filtered automatically
Previously we called rd_adc whenever we wanted a reading from the adc,
the main problem with this is that it's unfiltered.

This commit adds a method of filtering the adc by using timer0 as an
interrupt source. ISR(timer0) calls rd_adc repeatedly which has been
modified to update an internal buffer through a simple filter.

All code that previously used rd_adc must now use adc_get_values().
This function is a drop-in-replacement to rd_adc.